+ added /cgi-bin/zapto to http handling + added /xml/streaminfo to http handling + added Name2ServiceID to ChannelAndStationMapper
git-svn-id: https://svn.schrankmonster.de:8443/svn/YAPS/trunk@168 d4323bb5-e6f8-f049-b4c0-9a275b4df250

+++ YAPS_Processors/data structures/StationAndChannel.cs
@@ -11,7 +11,7 @@ namespace YAPS
         public String ChannelName;
         public Int32 ChannelNumber;           // legacy
         public String ChannelPictureURL;
-        public Int32 ServiceID;
+        public ushort ServiceID;
 
         public String MulticastIP;
         public String MulticastPort;
